# docs/swagger.yaml
definitions:
  dto.CloneCourseRequest:
    properties:
      author_id:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      image_path:
        type: string
      name:
        type: string
      parent_course_id:
        type: string
    required:
    - author_id
    - date_end
    - date_start
    - name
    - parent_course_id
    type: object
  dto.CreateCourse:
    properties:
      author_id:
        type: string
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      image_path:
        type: string
      name:
        type: string
    required:
    - author_id
    - date_end
    - date_start
    - name
    type: object
  dto.CreateCourseResponse:
    properties:
      id:
        type: string
    type: object
  dto.CreateModule:
    properties:
      date_start:
        type: string
      id:
        type: string
      name:
        type: string
      sequence_number:
        type: integer
    type: object
  dto.CreateModuleAttachmentRequest:
    properties:
      bucket:
        type: string
      file_name:
        type: string
      visible:
        type: boolean
    type: object
  dto.CreateModuleAttachmentResponse:
    properties:
      data:
        $ref: '#/definitions/models.ModuleAttachment'
      message:
        type: string
      status:
        type: boolean
    type: object
  dto.CreateModulesRequest:
    properties:
      course_id:
        type: string
      modules:
        items:
          $ref: '#/definitions/dto.CreateModule'
        type: array
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dto.GetModule:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.GetModuleResponse:
    properties:
      attachment:
        items:
          $ref: '#/definitions/models.ModuleAttachment'
        type: array
      module:
        $ref: '#/definitions/dto.GetModule'
      tasks:
        items:
          $ref: '#/definitions/dto.Task'
        type: array
    type: object
  dto.Task:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateCourseRequest:
    properties:
      date_end:
        type: string
      date_start:
        type: string
      description:
        type: string
      image_path:
        type: string
      name:
        type: string
    type: object
...
  models.ModuleAttachment:
    properties:
      bucket:
        type: string
      fileName:
        type: string
      id:
        type: string
      moduleID:
        type: string
      visible:
        type: boolean
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is an example API to demonstrate Swagger documentation generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms
  title: Example API
  version: "1.0"
paths:
  /courses:
    get:
      consumes:
      - application/json
      description: Получить список всех курсов с возможной фильтрацией
      parameters:
      - description: Название курса
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CreateCourseResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить все курсы
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Создание нового курса
      parameters:
      - description: Информация о курсе
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateCourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Создать курс
      tags:
      - courses
  /courses/{id}:
    delete:
      consumes:
      - application/json
      description: Удалить курс по ID
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Курс успешно удалён
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удалить курс
      tags:
      - courses
    get:
      consumes:
      - application/json
      description: Получить курс по ID
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить курс
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Обновить курс по ID
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: string
      - description: Информация о курсе
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновить курс
      tags:
      - courses
  /courses/{id}/clone:
    post:
      consumes:
      - application/json
      description: Клонировать курс по ID
      parameters:
      - description: ID родительского курса
        in: path
        name: id
        required: true
        type: string
      - description: Информация о новом курсе
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.CloneCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Клонировать курс
      tags:
      - courses
  /modules:
    post:
      consumes:
      - application/json
      description: Создает один или несколько модулей курса
      parameters:
      - description: Данные для создания модулей
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateModulesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Modules created successfully
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Создание модулей
      tags:
      - Modules
  /modules/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет модуль по его ID
      parameters:
      - description: UUID модуля
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Module not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удаление модуля
      tags:
      - Modules
    get:
      consumes:
      - application/json
      description: Получает модуль по его ID
      parameters:
      - description: UUID модуля
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные модуля
          schema:
            $ref: '#/definitions/dto.GetModuleResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Module not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получение модуля
      tags:
      - Modules
  /modules/{id}/attachments:
    post:
      consumes:
      - application/json
      description: Создает вложение для указанного модуля
      parameters:
      - description: UUID модуля
        in: path
        name: id
        required: true
        type: string
      - description: Данные вложения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateModuleAttachmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attachment created
          schema:
            $ref: '#/definitions/dto.CreateModuleAttachmentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Module not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Добавление вложения к модулю
      tags:
      - Modules
...
swagger: "2.0"